pyproject.toml
[project]
name = "ai-job-match-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "openai>=1.0.0",
    "pandas>=2.3.0",
    "playwright>=1.52.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.0",
]
